TST/REF: collect tests by method (#37403)
