+++ pandas/tests/arrays/categorical/test_indexing.py
@@ -91,6 +91,22 @@ class TestCategoricalIndexingWithFactor(TestCategorical):
         cat[1] = cat[0]
         assert cat[1] == (0, 1)
 
+    def test_setitem_listlike(self):
+
+        # GH#9469
+        # properly coerce the input indexers
+        np.random.seed(1)
+        c = Categorical(
+            np.random.randint(0, 5, size=150000).astype(np.int8)
+        ).add_categories([-1000])
+        indexer = np.array([100000]).astype(np.int64)
+        c[indexer] = -1000
+
+        # we are asserting the code result here
+        # which maps to the -1000 category
+        result = c.codes[np.array([100000]).astype(np.int64)]
+        tm.assert_numpy_array_equal(result, np.array([5], dtype="int8"))
+
 
 class TestCategoricalIndexing:
     def test_getitem_slice(self):
