+++ pandas/tests/frame/test_timezones.py
@@ -6,34 +6,12 @@ import pytest
 
 from pandas.core.dtypes.dtypes import DatetimeTZDtype
 
-import pandas as pd
 from pandas import DataFrame, Series
 import pandas._testing as tm
 from pandas.core.indexes.datetimes import date_range
 
 
 class TestDataFrameTimezones:
-    def test_frame_join_tzaware(self):
-        test1 = DataFrame(
-            np.zeros((6, 3)),
-            index=date_range(
-                "2012-11-15 00:00:00", periods=6, freq="100L", tz="US/Central"
-            ),
-        )
-        test2 = DataFrame(
-            np.zeros((3, 3)),
-            index=date_range(
-                "2012-11-15 00:00:00", periods=3, freq="250L", tz="US/Central"
-            ),
-            columns=range(3, 6),
-        )
-
-        result = test1.join(test2, how="outer")
-        ex_index = test1.index.union(test2.index)
-
-        tm.assert_index_equal(result.index, ex_index)
-        assert result.index.tz.zone == "US/Central"
-
     @pytest.mark.parametrize("tz", ["US/Eastern", "dateutil/US/Eastern"])
     def test_frame_no_datetime64_dtype(self, tz):
         # after GH#7822
@@ -94,11 +72,4 @@ class TestDataFrameTimezones:
         expected = DataFrame(
             np.arange(0, 5), index=date_range("20131027", periods=5, freq="1H", tz=tz)
         )
-        tm.assert_frame_equal(result, expected)
-
-    def test_constructor_data_aware_dtype_naive(self, tz_aware_fixture):
-        # GH 25843
-        tz = tz_aware_fixture
-        result = DataFrame({"d": [pd.Timestamp("2019", tz=tz)]}, dtype="datetime64[ns]")
-        expected = DataFrame({"d": [pd.Timestamp("2019")]})
         tm.assert_frame_equal(result, expected)
